Fix reconnect on E2 Setup Request
The E2Sim did not reconnect in case of no response or even on
response failure of the E2 setup request message.

This change fix this issue by trying to reconnect to the E2Term
every 10 seconds for 3 times. In case of no success, then the
E2Sim exits gracefully.