Move signal handler logic to stop the simulator
The logic to stop the E2Sim is handled in the e2sim base code and
is not flexible to support different use cases.

This change moves the signal handler logic to stop the e2sim from
the e2sim base code to the main application.
Thus, the reponsibility to catch signals is of the main application
which must call the shutdown logic of the e2sim base code to finish
the application gracefully.