Add sctp handover feature via O1 interface
The E2sim did not support to handover the insert-control loop
running on a given E2Term (sctp connetion) to another E2Term
(sctp connection).
This is required in situations that the insert-control loop is
above the 10ms threshold using the current E2Term, and there is another
E2Term that can meet that threshold if the sctp connection if handoff to
that E2Term.

This change implements this E2Term handover feature employing the O1
interface through RESTCONF that is intented to be called by SMO.