BIG CHANGE: Updates the architecture of the simulator
This change implements significant updates in the whole architecture
if the simulator. The main goal of these changes to speed up the
development of new features in the simulator, while maintaing
all the other features that were added previously.
The ideia is to implement a feature once, and keep it available
and runninng whenever required along with other new features
that are being implemented.
The simulator behavior that we want to achieve is something
similar to what gNodeBs do, by implementing several service
styles and possibly multiple service modules that can operate
along with one another.

Current change adds support to E2SM-RC Style 4 as follows:
  - Awaits for subscriptions to send reports of RCC-State-Change-To,
    RSRP, RSRQ, and SINR metrics every time a UE connects
    or disconnects from the RAN.

Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>