Implement E2SM-RC Control Style 3
This change implements Action ID 1 of E2SM-RC Control Style 3 to
enable handover control of UEs.
Some changes also have been introduces in other parts of the code
to accomodate the new changes with the older ones in the current
architecure.
Also, the default MNC value is changes from 01 to 001, as we
consider for now that all MNC are composed of 3 digits.

Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>